requirements.txt
Flask==3.0.3
flask-cors==5.0.0
ultralytics==8.3.47
Pillow==10.4.0
gunicorn==22.0.0